move poly. approximations to separate file
