+++ qualtran/bloqs/hamiltonian_simulation/hamiltonian_simulation_by_gqsp.py
@@ -11,13 +11,11 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-import bisect
 from functools import cached_property
 from typing import Dict, Set
 
 import cirq
 import numpy as np
-import scipy
 import sympy
 from attrs import field, frozen
 from numpy.typing import NDArray
@@ -25,40 +23,12 @@ from numpy.typing import NDArray
 from qualtran import Controlled, CtrlSpec, GateWithRegisters, Signature
 from qualtran.bloqs.basic_gates import SU2RotationGate
 from qualtran.bloqs.generalized_qsp import GeneralizedQSP
+from qualtran.bloqs.qsp.polynomial_approximations import (
+    approx_exp_cos_by_jacobi_anger,
+    degree_jacobi_anger_approximation,
+)
 from qualtran.bloqs.qubitization_walk_operator import QubitizationWalkOperator
 from qualtran.resource_counting.symbolic_counting_utils import SymbolicFloat, SymbolicInt
-
-
-def degree_jacobi_anger_approximation(t: float, precision: float) -> int:
-    r"""Degree of the Jacobi-Anger expansion of $e^{it\sin(\theta)}$ or $e^{it\cos(\theta)}$.
-
-    The Jacobi-Anger expansions are given by:
-
-    $$
-        e^{it\cos\theta} = \sum_{n = -\infty}^\infty i^n J_n(t) e^{in\theta}
-        e^{it\sin\theta} = \sum_{n = -\infty}^\infty J_n(t) e^{in\theta}
-    $$
-    where $J_n$ is the $n$-th Bessel function of the first kind.
-
-    We truncate the above series to the range $n \in [-d, d]$ such that $|J_{d+1}(t)| \le \epsilon$.
-
-    Args:
-        t: scale of the exponent in the function to approximate.
-        precision: $\epsilon$ in the above polynomial approximation
-
-    Returns:
-        Truncation degree $d$ as defined above.
-    """
-
-    def term_too_small(n: int) -> bool:
-        return np.isclose(scipy.special.jv(n, t), 0, atol=precision / 2)
-
-    d = 1
-    while not term_too_small(d):
-        d *= 2
-
-    # find the smallest `n` such that J_n(z) is too small
-    return bisect.bisect(range(d), True, key=term_too_small)
 
 
 @frozen
@@ -98,16 +68,14 @@ class HamiltonianSimulationByGQSP(GateWithRegisters):
                 + sympy.log(self.precision) / sympy.log(sympy.log(self.precision))
             )
         else:
-            return degree_jacobi_anger_approximation(self.t * self.alpha, self.precision)
+            return degree_jacobi_anger_approximation(self.t * self.alpha, precision=self.precision)
 
     @cached_property
     def approx_cos(self) -> NDArray[np.complex_]:
         r"""polynomial approximation for $$e^{i\theta} \mapsto e^{it\cos(\theta)}$$"""
         if self._parameterized_():
             raise ValueError(f"cannot compute `cos` approximation for parameterized Bloq {self}")
-        coeff_indices = np.arange(-self.degree, self.degree + 1)
-        approx_cos = 1j**coeff_indices * scipy.special.jv(coeff_indices, self.t * self.alpha)
-        return approx_cos
+        return approx_exp_cos_by_jacobi_anger(self.t * self.alpha, degree=self.degree)
 
     @cached_property
     def gqsp(self) -> GeneralizedQSP:
